refactor architecture: remove unused
