--- src/mimarsinan/pipelining/pipeline_steps/fine_tuning_step.py
@@ -1,34 +1,0 @@
-from mimarsinan.pipelining.pipeline_step import PipelineStep
-from mimarsinan.model_training.basic_trainer import BasicTrainer
-from mimarsinan.data_handling.data_loader_factory import DataLoaderFactory
-
-class FineTuningStep(PipelineStep):
-    def __init__(self, pipeline):
-        requires = ["model"]
-        promises = []
-        updates = ["model"]
-        clears = []
-        super().__init__(requires, promises, updates, clears, pipeline)
-
-        self.trainer = None
-
-    def validate(self):
-        return self.trainer.validate()
-
-    def process(self):
-        model = self.get_entry("model")
-
-        self.trainer = BasicTrainer(
-            model, 
-            self.pipeline.config['device'], 
-            DataLoaderFactory(self.pipeline.data_provider_factory),
-            self.pipeline.loss)
-        self.trainer.report_function = self.pipeline.reporter.report
-
-        self.trainer.train_until_target_accuracy(
-            self.pipeline.config['lr'] * 0.1, 
-            self.pipeline.config['tuner_epochs'],
-            target_accuracy=self.pipeline.get_target_metric(),
-            warmup_epochs=5)
-        
-        self.update_entry("model", model, 'torch_model')
