+++ src/mimarsinan/pipelining/pipeline_steps/soft_core_mapping_step.py
@@ -1,6 +1,5 @@
 from mimarsinan.pipelining.pipeline_step import PipelineStep
 
-from mimarsinan.pipelining.pipeline_steps.perceptron_fusion_step import FusedLinear
 from mimarsinan.mapping.ir_mapping import IRMapping
 from mimarsinan.mapping.ir_latency import IRLatency
 from mimarsinan.mapping.ir import NeuralCore
@@ -235,3 +234,26 @@ class SoftCoreMappingStep(PipelineStep):
         new_layer.bias.data = bias
 
         return new_layer
+
+
+class FusedLinear(nn.Module):
+    """Linear layer with bias fused into an extra weight column."""
+
+    def __init__(self, input_features, output_features):
+        super().__init__()
+        self.linear = nn.Linear(input_features + 1, output_features, bias=False)
+        self.weight = self.linear.weight
+        self.bias = self.linear.bias
+
+    def forward(self, x):
+        if len(x.shape) == 2:
+            x = x.unsqueeze(1)
+
+        batch_size, seq_len, _ = x.shape
+        bias_feature = torch.ones((batch_size, seq_len, 1), device=x.device)
+        x = torch.cat([x, bias_feature], dim=-1)
+        output = self.linear(x)
+
+        if len(output.shape) == 3 and output.shape[1] == 1:
+            output = output.squeeze(1)
+        return output
